add search

+++ TiktokBackend.Application/Commands/Auths/LoginCommand.cs
@@ -9,13 +9,14 @@ namespace TiktokBackend.Application.Commands.Auths
 {
     /// <summary>
     /// Lệnh đăng nhập, có thể sử dụng mật khẩu hoặc mã OTP.
-    /// Nếu Type == "phone" thì sử dụng OTP(Password = OTP), nếu Type == "email" thì sử dụng mật khẩu(Password = Password). 
+    /// Nếu Type == "phone" thì UserName = PhoneNumber, nếu Type == "email" thì UserName = UserName
+    /// Nếu Type == "phone" và IsOtp == true thì OTP(Password = OTP) ngược lại Password = Password, nếu Type == "email" thì sử dụng mật khẩu(Password = Password). 
     /// </summary>
-    public record LoginCommand(string Type,string UserName,string Password): IRequest<ServiceResponse<bool>>;
+    public record LoginCommand(string Type,string UserName,string Password,bool IsOtp = false): IRequest<ServiceResponse<bool>>;
     public class LoginCommandHandler : IRequestHandler<LoginCommand, ServiceResponse<bool>>
     {
         private readonly IUserRepository _userRepository;
-        private readonly IRedisService _redisService;
+        private readonly IOtpCacheService _redisService;
         private readonly IJwtService _jwtService;
         private readonly IUnitOfWork _unitOfWork;
         private readonly IUserRoleRepository _userRoleRepository;
@@ -23,7 +24,7 @@ namespace TiktokBackend.Application.Commands.Auths
         private readonly IUserContextService _userContextService;
         private readonly ICookieService _cookieService;
 
-        public LoginCommandHandler(IUserRepository userRepository,IRedisService redisService,IJwtService jwtService,
+        public LoginCommandHandler(IUserRepository userRepository,IOtpCacheService redisService,IJwtService jwtService,
             IUnitOfWork unitOfWork, IUserRoleRepository userRoleRepository, IUserTokenRepository userTokenRepository,
             IUserContextService userContextService, ICookieService cookieService)
         {
@@ -49,17 +50,26 @@ namespace TiktokBackend.Application.Commands.Auths
 
                 if (request.Type == "phone")
                 {
-                    string key = $"login:otp:{request.UserName}";
-                    var storedOtp = await _redisService.GetAsync(key);
-                    if (string.IsNullOrEmpty(storedOtp) || storedOtp != request.Password)
-                        return ServiceResponse<bool>.Fail("Mã xác thực không chính xác.");
+                    if (request.IsOtp)
+                    {
+                        string key = $"login:otp:{request.UserName}";
+                        var storedOtp = await _redisService.GetAsync(key);
+                        if (string.IsNullOrEmpty(storedOtp) || storedOtp != request.Password)
+                            return ServiceResponse<bool>.Fail("Mã xác thực không chính xác.");
 
-                    await _redisService.RemoveAsync(key);
-                    user = await _userRepository.GetUserByPhoneAsync(request.UserName);
+                        await _redisService.RemoveAsync(key);
+                        user = await _userRepository.GetUserByPhoneAsync(request.UserName);
+                    }
+                    else
+                    {
+                        user = await _userRepository.ValidateUserCasePhoneAsync(request.UserName,request.Password);
+                        if (user is null)
+                            return ServiceResponse<bool>.Fail("Mật khẩu hoặc tài khoản không chính xác!");
+                    }                 
                 }
                 else
                 {
-                    user = await _userRepository.ValidateUserAsync(request.UserName, request.Password);
+                    user = await _userRepository.ValidateUserCaseEmailAsync(request.UserName, request.Password);
                     if (user is null) 
                         return ServiceResponse<bool>.Fail("Mật khẩu hoặc tài khoản không chính xác!");
                 }
@@ -69,8 +79,9 @@ namespace TiktokBackend.Application.Commands.Auths
 
                 string ipAddress = _userContextService.GetIpAddress();
                 string userAgent = _userContextService.GetUserAgent();
+                string deviceId = _userContextService.GetDeviceId();
 
-                await _userTokenRepository.AddOrUpdateUserTokenAsync(user.Id, rftoken, ipAddress, userAgent);
+                await _userTokenRepository.AddOrUpdateUserTokenAsync(user.Id, rftoken, ipAddress, userAgent,deviceId);
                 await _unitOfWork.CommitAsync();
 
                 _cookieService.SetAccessToken(actoken);
